refactor: consolidate git helpers into lib/helpers/git.ts
Replace the single-function get-git-branch.ts with a general git.ts
module. Add getGitCommit() alongside getGitBranch() and use it in
trace-uploader.ts to replace its remaining open-coded spawnSync call,
removing the child_process import from that file entirely.
